🚨 Fix: 뒤로 가기로 이동 시 카메라 세팅 문제 해결
